Gitrepo Clean
A much needed repository clean moving older code into legacy folders so the currently used versions are clearer without removeing older versions for future reference
